requirement.txt
 async-generator==1.10
 attrs==22.2.0
 beautifulsoup4==4.11.2
 bs4==0.0.1
 certifi==2022.12.7
 cffi==1.15.1
 exceptiongroup==1.1.0
 h11==0.14.0
 idna==3.4
 install==1.3.5
 numpy==1.24.2
 outcome==1.2.0
 pandas==1.5.3
 pycparser==2.21
 PySocks==1.7.1
 python-dateutil==2.8.2
 pytz==2022.7.1
 PyYAML==6.0
 selenium==4.8.0
 six==1.16.0
 sniffio==1.3.0
 sortedcontainers==2.4.0
 soupsieve==2.3.2.post1
 trio==0.22.0
 trio-websocket==0.9.2
 tqdm
 urllib3==1.26.14
 wsproto==1.2.0
 PyPDF2